+++ Libraries/Jquery.php
@@ -30,79 +30,55 @@ class Jquery implements JavascriptInterface
             case '3.3.*':
             case '3.3.1':
             case '3.3.1.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.3.1.min.js',
-                    'integrity' => 'sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.3.1';
+                $integrity = 'sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=';
                 break;
 
             case '3.3.0':
             case '3.3.0.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.3.0.min.js',
-                    'integrity' => 'sha256-RTQy8VOmNlT6b2PIRur37p6JEBZUE7o8wPgMvu18MC4=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.3.0';
+                $integrity = 'sha256-RTQy8VOmNlT6b2PIRur37p6JEBZUE7o8wPgMvu18MC4=';
                 break;
 
             case '3.2.*':
             case '3.2.1':
             case '3.2.1.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.2.1.min.js',
-                    'integrity' => 'sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.2.1';
+                $integrity = 'sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=';
                 break;
 
             case '3.2.0':
             case '3.2.0.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.2.0.min.js',
-                    'integrity' => 'sha256-JAW99MJVpJBGcbzEuXk4Az05s/XyDdBomFqNlM3ic+I=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.2.0';
+                $integrity = 'sha256-JAW99MJVpJBGcbzEuXk4Az05s/XyDdBomFqNlM3ic+I=';
                 break;
 
             case '3.1.*':
             case '3.1.1':
             case '3.1.1.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.1.1.min.js',
-                    'integrity' => 'sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.1.1';
+                $integrity = 'sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8=';
                 break;
 
             case '3.1.0':
             case '3.1.0.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.1.0.min.js',
-                    'integrity' => 'sha256-cCueBR6CsyA4/9szpPfrX3s49M9vUU5BgtiJj06wt/s=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.1.0';
+                $integrity = 'sha256-cCueBR6CsyA4/9szpPfrX3s49M9vUU5BgtiJj06wt/s=';
                 break;
 
             case '3.0.*':
             case '3.0.0':
             case '3.0.0.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.0.0.min.js',
-                    'integrity' => 'sha256-JmvOoLtYsmqlsWxa7mDSLMwa6dZ9rrIdtrrVYRnDRH0=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.0.0';
+                $integrity = 'sha256-JmvOoLtYsmqlsWxa7mDSLMwa6dZ9rrIdtrrVYRnDRH0=';
                 break;
 
             case '2.*':
             case '2.2.*':
             case '2.2.4':
             case '2.2.4.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-2.2.4.min.js',
-                    'integrity' => 'sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '2.2.4';
+                $integrity = 'sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44=';
                 break;
 
             default:
@@ -110,8 +86,14 @@ class Jquery implements JavascriptInterface
                 break;
         }
 
-        //Removes defer attribute
-        $data['defer'] = false;
+        if (isset($version)) {
+            $data = array(
+                'src' => 'https://code.jquery.com/jquery-' . $version  . '.min.js',
+                'integrity' => $integrity,
+                'crossorigin' => 'anonymous',
+                'defer' => false,
+            );
+        }
 
         return $data;
     }