+++ Libraries/Jqueryslim.php
@@ -30,68 +30,47 @@ class Jqueryslim implements JavascriptInterface
             case '3.3.*':
             case '3.3.1':
             case '3.3.1.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.3.1.slim.min.js',
-                    'integrity' => 'sha256-3edrmyuQ0w65f8gfBsqowzjJe2iM6n0nKciPUp8y+7E=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.3.1';
+                $integrity = 'sha256-3edrmyuQ0w65f8gfBsqowzjJe2iM6n0nKciPUp8y+7E=';
                 break;
 
             case '3.3.0':
             case '3.3.0.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.3.0.slim.min.js',
-                    'integrity' => 'sha256-AMg3I7ya76OLPD9M+Mk7kqrA29HUn/FuGBfT/9Uf9ls=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.3.0';
+                $integrity = 'sha256-AMg3I7ya76OLPD9M+Mk7kqrA29HUn/FuGBfT/9Uf9ls=';
                 break;
 
             case '3.2.*':
             case '3.2.1':
             case '3.2.1.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.2.1.slim.min.js',
-                    'integrity' => 'sha256-k2WSCIexGzOj3Euiig+TlR8gA0EmPjuc79OEeY5L45g=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.2.1';
+                $integrity = 'sha256-k2WSCIexGzOj3Euiig+TlR8gA0EmPjuc79OEeY5L45g=';
                 break;
 
             case '3.2.0':
             case '3.2.0.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.2.0.slim.min.js',
-                    'integrity' => 'sha256-qLAv0kBAihcHZLI3fv3WITKeRsUX27hd6upBBa0MSow=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.2.0';
+                $integrity = 'sha256-qLAv0kBAihcHZLI3fv3WITKeRsUX27hd6upBBa0MSow=';
                 break;
 
             case '3.1.*':
             case '3.1.1':
             case '3.1.1.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.1.1.slim.min.js',
-                    'integrity' => 'sha256-/SIrNqv8h6QGKDuNoLGA4iret+kyesCkHGzVUUV0shc=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.1.1';
+                $integrity = 'sha256-/SIrNqv8h6QGKDuNoLGA4iret+kyesCkHGzVUUV0shc=';
                 break;
 
             case '3.1.0':
             case '3.1.0.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.1.0.slim.min.js',
-                    'integrity' => 'sha256-cRpWjoSOw5KcyIOaZNo4i6fZ9tKPhYYb6i5T9RSVJG8=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.1.0';
+                $integrity = 'sha256-cRpWjoSOw5KcyIOaZNo4i6fZ9tKPhYYb6i5T9RSVJG8=';
                 break;
 
             case '3.0.*':
             case '3.0.0':
             case '3.0.0.*':
-                $data = array(
-                    'src' => 'https://code.jquery.com/jquery-3.0.0.slim.min.js',
-                    'integrity' => 'sha256-Rf4BadfyCtsvHmO89BUZcbYvNNvZvOT08ALfEzvCsD0=',
-                    'crossorigin' => 'anonymous',
-                );
+                $version = '3.0.0';
+                $integrity = 'sha256-Rf4BadfyCtsvHmO89BUZcbYvNNvZvOT08ALfEzvCsD0=';
                 break;
 
             default:
@@ -99,8 +78,14 @@ class Jqueryslim implements JavascriptInterface
                 break;
         }
 
-        //Removes defer attribute
-        $data['defer'] = false;
+        if (isset($version)) {
+            $data = array(
+                'src' => 'https://code.jquery.com/jquery-' . $version  . '.slim.min.js',
+                'integrity' => $integrity,
+                'crossorigin' => 'anonymous',
+                'defer' => false,
+            );
+        }
 
         return $data;
     }